+++ CHANGELOG.md
@@ -8,7 +8,20 @@ If you like this project and find it useful, please consider giving it a star on
 
 <a href="https://www.buymeacoffee.com/luligugithub"><img src="https://matterbridge.io/assets/bmc-button.svg" alt="Buy me a coffee" width="120"></a>
 
-## [0.0.3] - Dev branch
+## [0.0.4] - 2026-06-08
+
+### Added
+
+- Add `Power Source` device type.
+- Add `Soil Sensor` device type.
+
+### Changed
+
+- Update `README.md`.
+
+<a href="https://www.buymeacoffee.com/luligugithub"><img src="https://matterbridge.io/assets/bmc-button.svg" alt="Buy me a coffee" width="80"></a>
+
+## [0.0.3] - 2026-06-07
 
 - Initial commit (repository reset).